+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Make functions more generic. Functions that expect a graph can accept an adjacency matrix. And functions that expect a VertexClustering (membership vector) can accept a sequence of integers.
+
 ## [0.1.5] 2025-02-19
 
 ### Changed